docs: Add docs to mdx
