+++ modules/preview-react/pill/stories/examples/WithCount.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+import {Pill} from '@workday/canvas-kit-preview-react/pill';
+
+export const ClickablePillWithCount = () => (
+  <Pill onClick={() => console.warn('clicked')}>
+    Shoes
+    <Pill.Count>30</Pill.Count>
+  </Pill>
+);
