+++ modules/preview-react/pill/stories/examples/test-avatar.png
